+++ www/main/settings/utils/index.js
@@ -1,1 +1,2 @@
 export { settingsNavObj, settingsObj } from './parse-settings';
+export { uploadImage } from './theme-bg-uploader';
