+++ www/main/viewer.js
@@ -17,7 +17,7 @@ const { remote } = require('electron');
 const { store, i18n } = require('electron').remote.require('./app');
 const slash = require('./js/slash');
 const core = require('./js/index');
-const themes = require('./configs/themes.json');
+// const themes = require('./configs/themes.json');
 
 const shortcuts = require('./js/keyboard-shortcuts/shortcuts');
 
@@ -200,9 +200,9 @@ global.platform.ipc.on('send-scroll', (event, pos) => {
 
 global.platform.ipc.on('update-settings', () => {
   prefs = store.get('userPrefs');
-  const themeKeys = themes.map(item => item.key);
-  $body.classList.remove(...themeKeys);
-  $body.classList.add(prefs.app.theme);
+  // const themeKeys = themes.map(item => item.key);
+  // $body.classList.remove(...themeKeys);
+  // $body.classList.add(prefs.app.theme);
   applyThemebg();
   core.menu.settings.applySettings(prefs);
   castToReceiver();
