+++ www/src/scss/_notifications.scss
@@ -44,7 +44,7 @@
   }
 }
 
-.dark-theme {
+.theme-dark-theme {
   .notifications-modal,
   .sync-code-modal {
     .tingle-modal-box__content {
